backend/requirements.txt
flask
flask-cors
supabase
python-dotenv
numpy
opencv-python
face-recognition
pickle-mixin